+++ src/main/java/jewellery/inventory/model/resource/PreciousStone.java
@@ -1,6 +1,7 @@
 package jewellery.inventory.model.resource;
 
 import jakarta.persistence.Entity;
+import java.math.BigDecimal;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
@@ -13,11 +14,11 @@ import lombok.experimental.SuperBuilder;
 @NoArgsConstructor
 public class PreciousStone extends Resource {
   private String color;
-  private Double carat;
+  private BigDecimal carat;
   private String cut;
   private String clarity;
-  private Double dimensionX;
-  private Double dimensionY;
-  private Double dimensionZ;
+  private BigDecimal dimensionX;
+  private BigDecimal dimensionY;
+  private BigDecimal dimensionZ;
   private String shape;
 }
